backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=4.2",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.0",
    "djangorestframework-simplejwt>=5.2",
    "channels>=4.0",
    "daphne>=4.0",
    "psycopg2-binary>=2.9",
    "django-filter>=23.0",
    "pillow>=10.0"
]
